+++ TO-DO.md
@@ -5,5 +5,6 @@
 - HTML export with collaps/expand projects
 - dark theme
 - "About" with a description and short help
-- Doe SyncSettings need targetRef?
-- split Store and Notifier
+- Does SyncSettings need targetRef?
+- Sync
+- Export/Import
